orcha01/requirements.txt
transformers
accelerate
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
scipy
tqdm
loguru
